Refactor files path
