+++ src/components/Content/Content.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/functions/toRem.scss';
+@import "/src/styles/functions/toRem.scss";
 
 .buttonContainer {
   display: flex;
