+++ src/components/MusicCatalogApp/MusicCatalogApp.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Content from "../Content/Content";
-import Header from "../Header/Header";
+import Content from "../Content";
+import Header from '../Header';
 
 const MusicCatalogApp = () => {
   return (
