+++ src/components/MusicCatalogList/MusicCatalogList.module.scss
@@ -1,2 +1,1 @@
-@import '../../styles/functions/toRem.scss';
-
+@import "/src/styles/functions/toRem.scss";
