+++ src/components/QuickModal/QuickModal.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Button from "../UI/Button/Button";
+import Button from "../UI/Button";
 import s from './QuickModal.module.scss';
 
 const QuickModal = ({setActive, children}) => {
