+++ src/components/QuickModal/QuickModal.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/functions/toRem.scss';
+@import "/src/styles/functions/toRem.scss";
 
 .container {
   position: fixed;
