+++ src/components/QuickModal/index.js
@@ -1,0 +1,1 @@
+export { default } from "./QuickModal";
