+++ src/components/Song/Song.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/functions/toRem.scss';
+@import "/src/styles/functions/toRem.scss";
 
 .container {
   // position: relative;
