+++ src/components/UI/Input/Input.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/functions/toRem.scss';
+@import "/src/styles/functions/toRem.scss";
 
 .error {
   font-size: toRem(12);
