+++ src/components/UI/Input/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Input";
