+++ src/index.js
@@ -2,10 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 import store from './store';
-// import '../styles/global';
 import App from './App';
 import './styles/global.scss';
-
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
